turned authenticate into an object
authenticate can now operate by itself since it's an object. sussed out
all of the requirements too! :D hooray!
